requirements.txt
qrcode[pil]
Pillow
python-dotenv